requirements.txt
django-taggit
django-autocomplete-light
django-debug-toolbar
djangorestframework
django-filter